Add fluid mobility and slip coefficient
Fluid mobility is defined as the ratio between a fluid density
and its dynamic viscosity.

In the issue description, we called it just mobility, but we already
have a mobility category with a different definition. So I've decided to
call it fluid mobility as also available in the issue description.

AGW-92

Add slip coefficient

Conceptually the slip coefficient is not a productivity index, because however both
have the same units and the productivity index is already being used everywhere.
So, we decided to stay simple and use it instead of defining a base unit that contains both.

ASIM-6432

Update changelog

AGW-92